Run the app during build phase
I'm not entirely sure why, but the initial run of the app in Heroku doesn't keep running it creates the config file then exits without error. When the app is run a 2nd time, it runs as expected. So added a "build" run of the app before the standard "start".